try to map context length exceeded errors based on status codes for providers
